In node context, catch uncaught exceptions and continue on
Still report the test failure with the error being thrown, but always
try to go as far as possible, instead of crashing entire test run at
the moment of async exception.

Most of our exceptions are ReferenceErrors, when some part of the test
fails and the failure wasn't guarded properly but only T.equal-ed or
T.assert-ed on. Example:

```
await get_body esc defer emails
T.assert emails?.length
T.equal emails[0].email, "test@example.com"
```

This will crash because even though emails?.length was checked, the test
does not exit when this check is made, but continues on. So `emails[0]` is
reached here and causes a crash.